refactor seeding script
